+++ src/lib.rs
@@ -31,6 +31,7 @@
 
 #![doc(html_logo_url = "https://libp2p.io/img/logo_small.png")]
 #![doc(html_favicon_url = "https://libp2p.io/img/favicon.png")]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 pub use bytes;
 pub use futures;
@@ -40,162 +41,89 @@ pub use libp2p_core::multihash;
 pub use multiaddr;
 
 #[cfg(feature = "autonat")]
-#[cfg_attr(docsrs, doc(cfg(feature = "autonat")))]
 #[doc(inline)]
 pub use libp2p_autonat as autonat;
 #[doc(inline)]
 pub use libp2p_core as core;
 #[cfg(feature = "dcutr")]
-#[cfg_attr(docsrs, doc(cfg(feature = "dcutr")))]
 #[doc(inline)]
 pub use libp2p_dcutr as dcutr;
 #[cfg(feature = "deflate")]
-#[cfg_attr(docsrs, doc(cfg(feature = "deflate")))]
 #[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
 #[doc(inline)]
 pub use libp2p_deflate as deflate;
-#[deprecated(
-    since = "0.49.0",
-    note = "The `dns-tokio` and `dns-async-std` features are deprecated. Use the new `dns` feature together with the `tokio` or `async-std` features."
-)]
-#[cfg(all(
-    any(feature = "dns-tokio", feature = "dns-async-std"),
-    not(feature = "dns")
-))]
-#[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
-#[cfg_attr(
-    docsrs,
-    doc(cfg(any(feature = "dns-tokio", feature = "dns-async-std")))
-)]
-pub mod dns {
-    #[doc(inline)]
-    pub use libp2p_dns::*;
-}
-
 #[cfg(feature = "dns")]
-#[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
 #[cfg_attr(docsrs, doc(cfg(feature = "dns")))]
-pub mod dns {
-    #[doc(inline)]
-    pub use libp2p_dns::*;
-}
+#[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
+#[doc(inline)]
+pub use libp2p_dns as dns;
 #[cfg(feature = "floodsub")]
-#[cfg_attr(docsrs, doc(cfg(feature = "floodsub")))]
 #[doc(inline)]
 pub use libp2p_floodsub as floodsub;
 #[cfg(feature = "gossipsub")]
-#[cfg_attr(docsrs, doc(cfg(feature = "gossipsub")))]
 #[cfg(not(target_os = "unknown"))]
 #[doc(inline)]
 pub use libp2p_gossipsub as gossipsub;
 #[cfg(feature = "identify")]
-#[cfg_attr(docsrs, doc(cfg(feature = "identify")))]
 #[doc(inline)]
 pub use libp2p_identify as identify;
 #[cfg(feature = "kad")]
-#[cfg_attr(docsrs, doc(cfg(feature = "kad")))]
 #[doc(inline)]
 pub use libp2p_kad as kad;
-#[deprecated(
-    since = "0.49.0",
-    note = "The `mdns-tokio` and `mdns-async-io` features are deprecated. Use the new `mdns` feature together with the `tokio` or `async-std` features."
-)]
-#[cfg(all(
-    any(feature = "mdns-async-io", feature = "mdns-tokio"),
-    not(feature = "mdns")
-))]
-#[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
-#[cfg_attr(
-    docsrs,
-    doc(cfg(any(feature = "mdns-tokio", feature = "mdns-async-io")))
-)]
-pub mod mdns {
-    #[doc(inline)]
-    pub use libp2p_mdns::*;
-}
-
 #[cfg(feature = "mdns")]
 #[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
 #[cfg_attr(docsrs, doc(cfg(feature = "mdns")))]
-pub mod mdns {
-    #[doc(inline)]
-    pub use libp2p_mdns::*;
-}
+#[doc(inline)]
+pub use libp2p_mdns as mdns;
 #[cfg(feature = "metrics")]
-#[cfg_attr(docsrs, doc(cfg(feature = "metrics")))]
 #[doc(inline)]
 pub use libp2p_metrics as metrics;
 #[cfg(feature = "mplex")]
-#[cfg_attr(docsrs, doc(cfg(feature = "mplex")))]
 #[doc(inline)]
 pub use libp2p_mplex as mplex;
 #[cfg(feature = "noise")]
-#[cfg_attr(docsrs, doc(cfg(feature = "noise")))]
 #[doc(inline)]
 pub use libp2p_noise as noise;
 #[cfg(feature = "ping")]
-#[cfg_attr(docsrs, doc(cfg(feature = "ping")))]
 #[doc(inline)]
 pub use libp2p_ping as ping;
 #[cfg(feature = "plaintext")]
-#[cfg_attr(docsrs, doc(cfg(feature = "plaintext")))]
 #[doc(inline)]
 pub use libp2p_plaintext as plaintext;
 #[cfg(feature = "pnet")]
-#[cfg_attr(docsrs, doc(cfg(feature = "pnet")))]
 #[doc(inline)]
 pub use libp2p_pnet as pnet;
 #[cfg(feature = "relay")]
-#[cfg_attr(docsrs, doc(cfg(feature = "relay")))]
 #[doc(inline)]
 pub use libp2p_relay as relay;
 #[cfg(feature = "rendezvous")]
-#[cfg_attr(docsrs, doc(cfg(feature = "rendezvous")))]
 #[doc(inline)]
 pub use libp2p_rendezvous as rendezvous;
 #[cfg(feature = "request-response")]
-#[cfg_attr(docsrs, doc(cfg(feature = "request-response")))]
 #[doc(inline)]
 pub use libp2p_request_response as request_response;
 #[doc(inline)]
 pub use libp2p_swarm as swarm;
-#[deprecated(
-    since = "0.49.0",
-    note = "The `tcp-tokio` and `tcp-async-io` features are deprecated. Use the new `tcp` feature together with the `tokio` or `async-std` features."
-)]
-#[cfg(all(
-    any(feature = "tcp-tokio", feature = "tcp-async-io"),
-    not(feature = "tcp")
-))]
-#[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
-#[cfg_attr(docsrs, doc(cfg(any(feature = "tcp-tokio", feature = "tcp-async-io"))))]
-pub mod tcp {
-    #[doc(inline)]
-    pub use libp2p_tcp::*;
-}
-
 #[cfg(feature = "tcp")]
 #[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
 #[cfg_attr(docsrs, doc(cfg(feature = "tcp")))]
-pub mod tcp {
-    #[doc(inline)]
-    pub use libp2p_tcp::*;
-}
+#[doc(inline)]
+pub use libp2p_tcp as tcp;
+#[cfg(feature = "tls")]
+#[cfg_attr(docsrs, doc(cfg(feature = "tls")))]
+#[doc(inline)]
+pub use libp2p_tls as tls;
 #[cfg(feature = "uds")]
-#[cfg_attr(docsrs, doc(cfg(feature = "uds")))]
 #[doc(inline)]
 pub use libp2p_uds as uds;
 #[cfg(feature = "wasm-ext")]
-#[cfg_attr(docsrs, doc(cfg(feature = "wasm-ext")))]
 #[doc(inline)]
 pub use libp2p_wasm_ext as wasm_ext;
 #[cfg(feature = "websocket")]
-#[cfg_attr(docsrs, doc(cfg(feature = "websocket")))]
 #[cfg(not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")))]
 #[doc(inline)]
 pub use libp2p_websocket as websocket;
 #[cfg(feature = "yamux")]
-#[cfg_attr(docsrs, doc(cfg(feature = "yamux")))]
 #[doc(inline)]
 pub use libp2p_yamux as yamux;
 
@@ -242,20 +170,6 @@ pub use libp2p_swarm_derive::NetworkBehaviour;
     feature = "yamux"
 ))]
 #[cfg_attr(
-    docsrs,
-    doc(cfg(all(
-        not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")),
-        any(
-            all(feature = "tcp-async-io", feature = "dns-async-std"),
-            all(feature = "tcp", feature = "dns", feature = "async-std")
-        ),
-        feature = "websocket",
-        feature = "noise",
-        feature = "mplex",
-        feature = "yamux"
-    )))
-)]
-#[cfg_attr(
     all(
         any(feature = "tcp-async-io", feature = "dns-async-std"),
         not(feature = "async-std")
@@ -269,13 +183,13 @@ pub async fn development_transport(
     keypair: identity::Keypair,
 ) -> std::io::Result<core::transport::Boxed<(PeerId, core::muxing::StreamMuxerBox)>> {
     let transport = {
-        let dns_tcp = dns::DnsConfig::system(tcp::TcpTransport::new(
-            tcp::GenTcpConfig::new().nodelay(true),
+        let dns_tcp = dns::DnsConfig::system(tcp::async_io::Transport::new(
+            tcp::Config::new().nodelay(true),
         ))
         .await?;
         let ws_dns_tcp = websocket::WsConfig::new(
-            dns::DnsConfig::system(tcp::TcpTransport::new(
-                tcp::GenTcpConfig::new().nodelay(true),
+            dns::DnsConfig::system(tcp::async_io::Transport::new(
+                tcp::Config::new().nodelay(true),
             ))
             .await?,
         );
@@ -316,20 +230,6 @@ pub async fn development_transport(
     feature = "yamux"
 ))]
 #[cfg_attr(
-    docsrs,
-    doc(cfg(all(
-        not(any(target_os = "emscripten", target_os = "wasi", target_os = "unknown")),
-        any(
-            all(feature = "tcp-tokio", feature = "dns-tokio"),
-            all(feature = "tcp", feature = "dns", feature = "tokio")
-        ),
-        feature = "websocket",
-        feature = "noise",
-        feature = "mplex",
-        feature = "yamux"
-    )))
-)]
-#[cfg_attr(
     all(
         any(feature = "tcp-tokio", feature = "dns-tokio"),
         not(feature = "tokio")
@@ -343,11 +243,11 @@ pub fn tokio_development_transport(
     keypair: identity::Keypair,
 ) -> std::io::Result<core::transport::Boxed<(PeerId, core::muxing::StreamMuxerBox)>> {
     let transport = {
-        let dns_tcp = dns::TokioDnsConfig::system(tcp::TokioTcpTransport::new(
-            tcp::GenTcpConfig::new().nodelay(true),
+        let dns_tcp = dns::TokioDnsConfig::system(tcp::tokio::Transport::new(
+            tcp::Config::new().nodelay(true),
         ))?;
         let ws_dns_tcp = websocket::WsConfig::new(dns::TokioDnsConfig::system(
-            tcp::TokioTcpTransport::new(tcp::GenTcpConfig::new().nodelay(true)),
+            tcp::tokio::Transport::new(tcp::Config::new().nodelay(true)),
         )?);
         dns_tcp.or_transport(ws_dns_tcp)
     };